+++ src/components/Share/Layout/index.ts
@@ -1,0 +1,1 @@
+import ShareLayout from './ShareLayout';
