+++ src/components/Share/index.ts
@@ -1,0 +1,2 @@
+import FlowerShareGuide from './FlowerShareGuide';
+import PicmeVoteShareGuide from './PicmeVoteShareGuide';
